6309433 gcc and dr don't get along 6309468 gcc and cvc,cvcredir don't get along 6309745 gcc and tazmo envctrl don't get along 6309752 gcc and littleneck pcf8574 don't get along 6309967 gcc and esp don't get along

+++ usr/src/uts/sun4u/ngdr/io/dr_io.c
@@ -188,7 +188,7 @@ dr_check_io_refs(dr_handle_t *hp, dr_common_unit_t **devlist, int devnum)
 			reftotal += ref;
 		} else {
 			PR_IO("%s: NO dip for id (0x%x)\n",
-				f, (uint_t)ip->sbi_cm.sbdev_id);
+				f, (uint_t)(uintptr_t)ip->sbi_cm.sbdev_id);
 		}
 	}
